put tempdir creation func in helpers
